requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-OAuthlib==0.9.4
Flask-PyMongo==0.5.1
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauth2==1.9.0.post1
oauthlib==2.0.2
pymongo==3.5.1
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
Werkzeug==0.12.2
python-google-places==1.4.1
yelpapi==2.2.2
flask-sqlalchemy==2.3.1
Flask-Migrate==2.1.1
google-auth==1.4.1
PyJWT==1.5.3
psycopg2==2.7.4
PyYAML==3.12
